vijil_dome/tutorials/MCP/requirements.txt
vijil_dome=1.4.1
python-dotenv=1.1.1
langchain=0.3.23
langchain_openai=0.3.14
langgraph==1.0.3

